HomePage's log updater transfered to subcomponent
Fixed SingleResponsibilityPrinciple for HomePage to subcomponenst
